feat: Add twMerge function to Stack component
The code changes in this commit add the `twMerge` function from the `tailwind-merge` library to the Stack component. This function allows for the merging of Tailwind CSS classes, providing a convenient way to combine multiple class names into a single string. By using `twMerge`, we can easily add additional CSS classes to the Stack component without overriding the existing ones.